refactor: Separate InstallationStep in two part (web and mobile)
